Introduce "simple custimazation" section in stylesheet
We're using a couple of cascading variables so even unexperienced users
should be able to do some customization.  This is much easier for us
than adding respective configuration options and supporting these with
more complex selectors, what would currently be required, since
xhstyles.css is added to `$hjs` so early, that we can't inject our own
rules above.

Still a small improvement that can come in handy for users.