Employ ts-check
We declare var and parameter types, and let tsc infer the rest.  To
satisfy the type checker, we add type casts where we can be sure that
the expression has either the type or `null` (the latter will blow up
soon anyway); for the couple of other cases, we actually throw if our
assertion fails.

This is certainly not fool-proof, but should be good enough for our
purposes, and feels much better than completely ignoring type issues,
and to eschew auto-completion.

We also run tsc in CI.