Create a set of thumbnails for `srcset`
Like mentioned before[1], we have no idea how big the thumbnails should
be.  To cater to that, we can no longer use a single thumbnail, but
rather need a set of thumbnails which we pass as `srcset`, so the
browser can fetch the best fitting thumbnail.

Since the thumbnail creation is now more expensive, we create the
thumbnails when saving the gallery in the back-end instead of every
plugin call processing.  When the gallery is going to be rendered, we
just look up all potential thumbnails in the cache and use these.

The whole implementation needs improvements and clean-up; will do this
in follow-up commits.

[1] <caab9c3f2e6f48bbafb811bd6ecef7c972ce605b>