Improve image size hinting for browsers
As is, we don't give the browser sufficient hint about the size of the
thumbnails; we just tell it that they would be roughly a third of the
viewport.  We can't do better, so we need to defer the image loading so
we can set `sized="auto"`.  However, that trips up the layout since the
browser assumes 300x150 px images, so we need to specify the intrinsic
width and height of the images.

As a quick solution we use `ImageFinder::size()`, which wraps
`getimagesize()`, so doesn't know about the Exif orientation.  We'll
fix that soon.