Cater to Exif orientation for "quick" size detection
`getimagesize()` doesn't know about the Exif orientation, so we need to
cater to that ourselves.

This solution works, but is not optimal regarding code duplication and
performance.  It might be reasonable to store the width and height in
the gallery XML, and to update the attributes when saving a gallery.
Let's see.