Show dialog about slow saving of galleries
For large galleries, saving takes quite a while, especially if the
thumbnails need to be created.  While there is not much we can do about
that, we can at least present a dialog with a suitable message and an
indeterminate progress bar.  A documentation only solution is likely to
be overlooked anyway.