Unify thumbnail handling
So far we created the set of thumbnails for the lightbox frontend
completely separate from the main thumbnail.  That doesn't make sense;
instead we can just use the smallest thumbnail from the set.  That is
sufficiently large for the Fotorama thumbsnails (which are tiny), and
likely good enough for the lightbox gallery; browsers which don't
support srcset should be very rare, and users of such browsers are
likely accustomed to a suboptimal surf experience.