Fix handling of external images
We need to cater to external images when initializing the gallery
editor, and also when saving.

But most importantly, we bring back the inadvertently removed special
thumbnail, which is important for privacy reasons; we should avoid any
connections to external websites until the visitor agrees.

To avoid scaling issues with external.jpg, we replace it with
external.svg, which is also suitable for `object-fit:fill`.

All in all, adding support for external images was a stupid idea for
many reasons, and we should consider to deprecate that.