Don't show images outside of image folder
We can't create thumbnails for these images, so large images would need
to be loaded for the gallery view what is really bad for performance;
the whole point of the thumbnails is to have bearable performance even
with slow connections.

Thus we no longer support images outside of the image folder in the
gallery front-end, although we still allow these in the back-end.