Ensure that progress dialog is closed when navigating
As is, when navigating back to the editor via the browser history after
the gallery has been saved, the save progress dialog is still shown.
That is highly confusing, so we close the dialog when the `pagehide`
event is fired (`unload` seems more appropriate to me, but the event
is deprecated).