Introduce assert()
For some reason, JavaScript in the browser doesn't have a built-in
assertion functionality, so we roll our own.  We replace all narrowing
"type casts" and all manually "inlined" assertions with `assert()`.
The latter is an obvious improvement, since `assert()` is more concise;
the former is also an improvement at least for development, since we're
informed right away about issues, instead of stumbling upon these
sometime later.

The obvious drawback is that we need an external module so that we can
inform typescript about the assertion semantics of the function.  This
is not a real issue during development, but for production we likely
want to inline the function declaration, and probably even get rid of
the assertions in the minified code.  We still need to figure out how
to best do this.