Replace arrow function expressions with function expressions
We almost always need the expression body syntax anyway, so arrow
function expression don't buy us much regarding the code size.  And
after all, JavaScript is somewhat confusing, but being able to
*sometimes* avoid some confusion with a newer construct actually adds
to the already existing confusion, so we stick with classic function
expression.  This also has the benefit that extracting or inlining is
more straight forward.