Use var instead of const/let
We believe that introducing block scope to JavaScript might not have
been the best idea: if a function is small, block scope is of small or
no benefit; if a function is big, it is likely better to split it.

Now there is an issue regarding TypeScript: type assertions for a const
are (mostly!) propagated to nested functions, but that doesn't happen
for vars.  So to satisfy tsc, we need to fetch some elements again, but
that might even be a good thing, since too many variables in outer
scopes are hard to track and maintain, and it might drive us even more
strongly to improve the code mess we've created.