Fix #40: Drag&drop does not work in Firefox
Firefox apparently does not trigger `dragstart` on `<input type=image>`
for whatever reason.  Thus we change that to `<img>` and add `tabindex`
so the images can still be focused with the keyboard.  Unfortunately,
Firefox shows the drag image with the natural dimensions (unlike Chrome
which shows the resized thumbs), so we do the resizing ourselves in a
`canvas`.  Since Chrome does not show drag images for elements not part
of the DOM, we insert the canvas into the DOM.  We can not hide it
there, since Chrome still wouldn't display the drag image, so we place
it offscreen.  That might not be the best idea wrt assistive technology
but in this case drag & drop is unlikely to be used anyway.

Apparently not much has changed from the days of the browser wars: it
is futile to rely on "standards" and especially "baseline" is utmost
nonsense, given that even the latest versions of Chrome and Firefox can
not agree how to work.  But at least we're grateful for the JavaScript
language improvements (ES6+) – which help to kill the Web faster.