Define setLink() dynamically
There is actually no need for editorhook.php which is just a clumsy
solution.  And since we define `setLink()` dynamically, we can as well
bind the required parameters up-front, so that we don't need the
"global" state, which would only ever work for a single filebrowser,
i.e. we no longer require a true modal filebrowser.