Support gallery viewer on older browsers
There is not much point in only supporting the latest and greatest
browsers, since neither we, nor the scripts we're using require ES6 and
alike (well, simple-lightbox would, but we just use the legacy variant,
which somewhat unfortunately is transpiled and apparently contains some
polyfills, but that's live[1]).

[1] It wouldn't have to be that way, but that would require that
front-end JavaScript developers actually understand JavaScript –
unlikely to ever happen. :(