Fix #45: Creating AVIF thumbnail may cause fatal error
We employ `getimagesize()` to get the MIME type of the image upfront –
if that fails, we don't even try to create the thumbnails.  While we
generally don't like `getimagesize()` for its many quirks, in this case
using it seems to be justified, and if there are no thumbnails, the
galleries will work anyway (by using the original image).