Avoid pseudo-dialogs at all
Pseudo-dialogs are not accessible.  While it is possible to enhance
these for accessibility[1], that's practically impossible with the
filebrowser (which is still not accessible).  Thus, we open the
filebrowser in a new window instead – that's neither as elegant as a
dialog, nor is it without accessibility problems, but it's much better
than a simple overlay, and much simpler.

[1] <https://bitsofco.de/accessible-modal-dialog/>