Back to nested functions
In 58140870c039636648ee9613176ffe167b91e145 we argued that an object
would yield a more maintainable code structure, but in the meantime we
seriously doubt that.  A particular issue is the occasional need to
call `.removeEventListener()`, and if we need to remove a method, we
would need to store the pre-bound method on the object – seriously
messy.  The alternative would be to implement `.handleEvent` and to
misuse this mechanism to bind `this`, but that implies event delegation
on steroids.

And generally, the dynamic binding of `this` is a mess, we better it
avoid altogether (as far as possible).  We need to be careful regarding
free variables in closures, but with some discipline this should work
out.  And after all, nesting functions to provide locality and to hide
information is what Schemers are successfully doing for decades.

A nice side-effect of using functions is that minification is way more
aggressive, so we can save some bandwidth.  A drawback is that multiple
widgets on the same page now require more memory, but we rarely have
many, and compared to the DOM's memory usage, that is just noise.