Fix drag and drop
When we switched to a more generally supported drag format[1], we did
not cater to other drag operations than our image reordering.  So if
somebody drags text from elsewhere onto the list (e.g. to the caption
or description fields), that is no longer possible; instead we would
treat that as image reordering, and almost certainly fail.

The obvious solution is to attach the drag/drop event listeners only if
a drag operation has been started on a thumbnail, and to remove the
listeners on dragend.

Now that we need the other handlers only in `onDragStart()`, we move
them into that function.  And we can now drop the types check, since
we can be sure that the drag operation has been initiated with the
proper type.

[1] <3037950b984b07644f9d04f2ac9af90ce58eac0f>